Move \begin handling to environment.js
